moved utility functions into JoySender.h for easy inclusion
